fix[Refactor] : Refactor package for game DTO
